Rename website category
