+++ public/sitemap.txt
@@ -14,9 +14,9 @@ https://bradyanderson.tech/blog/retro-gaming/replacing-game-boy-color-speaker
 https://bradyanderson.tech/blog/retro-gaming/game-gear-repair-part-1
 https://bradyanderson.tech/blog/retro-gaming/game-gear-repair-part-2
 https://bradyanderson.tech/blog/retro-gaming/game-gear-repair-part-3
-https://bradyanderson.tech/blog/bradyanderson.tech
-https://bradyanderson.tech/blog/bradyanderson.tech/getting-started
-https://bradyanderson.tech/blog/bradyanderson.tech/custom-hover-domain-for-heroku-app
-https://bradyanderson.tech/blog/bradyanderson.tech/moving-to-aws
-https://bradyanderson.tech/blog/bradyanderson.tech/setting-up-image-storage
-https://bradyanderson.tech/blog/bradyanderson.tech/dynamically-updating-html-title
+https://bradyanderson.tech/blog/website-updates
+https://bradyanderson.tech/blog/website-updates/getting-started
+https://bradyanderson.tech/blog/website-updates/custom-hover-domain-for-heroku-app
+https://bradyanderson.tech/blog/website-updates/moving-to-aws
+https://bradyanderson.tech/blog/website-updates/setting-up-image-storage
+https://bradyanderson.tech/blog/website-updates/dynamically-updating-html-title
